requirements.txt
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.11.0
autopep8==1.6.0
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
distro==1.9.0
Django==6.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.47.0
google-auth-httplib2==0.2.1
google-genai==1.57.0
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.62.3
gunicorn==21.2.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.11
jiter==0.12.0
lxml==6.0.2
mysql-connector-python==9.4.0
openai==2.14.0
packaging==25.0
pdfminer.six==20251107
pillow==12.1.0
proto-plus==1.26.1
protobuf==4.25.8
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.8.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.3.0
pyparsing==3.2.5
PyPDF2==3.0.1
python-docx==1.1.0
python-dotenv==1.2.1
pytz==2021.3
requests==2.31.0
rsa==4.9.1
setuptools==80.9.0
sniffio==1.3.1
sqlparse==0.5.5
tenacity==9.1.2
toml==0.10.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2021.5
uritemplate==4.2.0
urllib3==2.5.0
websockets==15.0.1
whitenoise==6.6.0
